Update aseprite-rgba to support linked cels.
Add dirty flag and loop types.

Fix a few template definitions.

Fix aserprite convert on gcc.

Fix __FUNCTION__ on gcc.

Fix frame time in renderable.

Normalize uinit32_t instaed of size_t for fixed sized counters.

+++ Farquaad/include/Farquaad/Components/Physics.hpp
@@ -85,5 +85,5 @@ public:
     CoollidingSet collidingWithSet;
     bool isColliding = false;
     bool isBullet = false;
-    size_t collisionCount = 0;
+    uint32_t collisionCount = 0;
 };
